requirements.txt
Flask==3.1.2
pymongo==4.15.4
requests==2.32.5
gunicorn==23.0.0
python-dotenv==1.2.1
